breaks App into smaller components
- created Loader component from App loading implementation
- created Main component from App main implementation
- moved Header and Footer components to components/Blocks folder
- updated App to use new components
- fixed unlimited card height bug
