+++ src/App/App.css
@@ -2,92 +2,3 @@
   font-family: 'Star Jedi';
   src: url(../assets/fonts/starjedi.ttf);
 }
-
-#temp-page {
-  max-width: 100%;
-  max-height: 100vh;
-  height: 100%;
-}
-
-#main {
-  background-image: url('../assets/img/bg.jpg');
-  padding: 15px 0;
-}
-
-.page-loader {
-  text-align: center;
-  vertical-align: center;
-  font-family: 'Star Jedi';
-  font-size: 10vh;
-  padding: 25vh 10vh;
-}
-
-@keyframes flickerAnimation {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-@-o-keyframes flickerAnimation {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-@-moz-keyframes flickerAnimation {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-@-webkit-keyframes flickerAnimation {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-.animate-flicker {
-  -webkit-animation: flickerAnimation 1s infinite;
-  -moz-animation: flickerAnimation 1s infinite;
-  -o-animation: flickerAnimation 1s infinite;
-  animation: flickerAnimation 1s infinite;
-}
-
-.white-backdrop {
-  background-color: rgba(240, 240, 240);
-  border-color: gray;
-}
-
-#main > .white-backdrop {
-  min-height: 67vh;
-  background-color: rgba(255, 255, 255, 0.8);
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
