+++ src/App/App.js
@@ -1,16 +1,14 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 
-import Header from '../components/Header';
-import SearchBox from '../components/SearchBox';
-import CardList from '../components/Cardlist';
-import Dialog from '../components/Dialog';
-import Footer from '../components/Footer';
-import { getStats, getFilms } from '../utils/functions.utils';
+import Footer from '../components/Blocks/Footer';
+import Header from '../components/Blocks/Header';
+import Main from '../components/Blocks/Main';
+import Loader from '../components/Loader';
 import './App.css';
 
 const App = ({
-  characters,
+  isCharactersLoading,
   loadCharacters,
   loadFilms,
   loadPlanets,
@@ -18,12 +16,6 @@ const App = ({
   loadStarships,
   loadVehicles,
 }) => {
-  const isLoadingCharacters = !characters.length;
-  const [searchField, setSearchField] = useState('');
-  const [clickedCard, setClickedCard] = useState({});
-  const [showModal, setShowModal] = useState(false);
-  const [characterDetails, setCharacterDetails] = useState({ race: {}, planet: {}, movies: '' });
-
   useEffect(() => {
     const fetchData = async () => {
       await loadCharacters();
@@ -37,65 +29,19 @@ const App = ({
     fetchData();
   }, [loadCharacters, loadFilms, loadPlanets, loadSpecies, loadStarships, loadVehicles]);
 
-  const onSearchChange = (event) => {
-    setSearchField(event.target.value);
-  };
-
-  const openModal = async (event) => {
-    setShowModal(true);
-    setClickedCard(characters[event.target.id]);
-    const race = await getStats(characters[event.target.id].species);
-    const planet = await getStats(characters[event.target.id].homeworld);
-    const movies = await getFilms(characters[event.target.id].films);
-    setCharacterDetails({
-      race,
-      planet,
-      movies,
-    });
-  };
-
-  const onCloseClick = () => {
-    setCharacterDetails({
-      race: {},
-      planet: {},
-      movies: '',
-    });
-    setClickedCard({});
-    setShowModal(false);
-  };
-
-  const filteredCharacters = characters.filter((character) =>
-    character.name.toLowerCase().includes(searchField.toLowerCase())
-  );
-
-  return isLoadingCharacters ? (
-    <div id="temp-page">
-      <div className="page-loader animate-flicker">Please wait...</div>
-    </div>
+  return isCharactersLoading ? (
+    <Loader />
   ) : (
     <>
       <Header />
-      <div id="main">
-        <SearchBox searchChange={onSearchChange} />
-        <CardList characters={filteredCharacters} openModal={openModal} />
-        ''
-        {showModal && (
-          <Dialog
-            character={clickedCard}
-            race={characterDetails.race}
-            planet={characterDetails.planet}
-            movies={characterDetails.movies}
-            onCloseClick={onCloseClick}
-          />
-        )}
-      </div>
+      <Main />
       <Footer />
     </>
   );
 };
 
 App.propTypes = {
-  characters: PropTypes.array.isRequired,
+  isCharactersLoading: PropTypes.bool.isRequired,
   loadCharacters: PropTypes.func.isRequired,
   loadFilms: PropTypes.func.isRequired,
   loadPlanets: PropTypes.func.isRequired,
