+++ src/components/Blocks/Header/header.styles.css
@@ -12,4 +12,3 @@
   max-height: 100vh;
   height: 100%;
 }
-
