+++ src/components/Blocks/Main/main.styles.css
@@ -1,0 +1,4 @@
+#main {
+  background-image: url('../../../assets/img/bg.jpg');
+  padding: 15px 0;
+}
