+++ src/components/Dialog/dialog.styles.css
@@ -29,7 +29,16 @@
   text-align: center;
 }
 
-.loader {
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+.dialog-loader {
   border: 2vw solid #f3f3f3;
   border-top: 2vw solid goldenrod;
   border-radius: 50%;
@@ -82,47 +91,47 @@
 }
 
 @media only screen and (max-width: 735px) {
-    .dialog-content,
-    .dialog-load {
-      width: 80vw;
-      margin: 2vh auto;
-    }
-  
-    .stats {
-      justify-self: start;
-      font-size: 2.8vw;
-    }
+  .dialog-content,
+  .dialog-load {
+    width: 80vw;
+    margin: 2vh auto;
   }
-  
-  @media only screen and (max-width: 516px) {
-    .dialog-content,
-    .dialog-load {
-      width: 60vh;
-      height: auto;
-      margin: 5vh auto;
-    }
-  
-    .stats-list {
-      grid-template-columns: auto;
-      grid-template-rows: auto;
-      padding: 10px;
-      justify-items: center;
-      grid-row-gap: 5px;
-    }
-  
-    .char-name {
-      font-size: 6vw;
-      margin-bottom: 2vw;
-    }
-  
-    .portrait {
-      grid-row: 2 / span 7;
-      width: 60%;
-      margin-bottom: 2vw;
-    }
-  
-    .stats {
-      justify-self: start;
-      font-size: 2.1vh;
-    }
+
+  .stats {
+    justify-self: start;
+    font-size: 2.8vw;
   }
+}
+
+@media only screen and (max-width: 516px) {
+  .dialog-content,
+  .dialog-load {
+    width: 60vh;
+    height: auto;
+    margin: 5vh auto;
+  }
+
+  .stats-list {
+    grid-template-columns: auto;
+    grid-template-rows: auto;
+    padding: 10px;
+    justify-items: center;
+    grid-row-gap: 5px;
+  }
+
+  .char-name {
+    font-size: 6vw;
+    margin-bottom: 2vw;
+  }
+
+  .portrait {
+    grid-row: 2 / span 7;
+    width: 60%;
+    margin-bottom: 2vw;
+  }
+
+  .stats {
+    justify-self: start;
+    font-size: 2.1vh;
+  }
+}
