+++ src/components/Loader/index.js
@@ -1,0 +1,3 @@
+import Loader from './loader.component';
+
+export default Loader;
