+++ src/components/Loader/loader.component.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './loader.styles.css';
+
+const Loader = () => (
+  <div className="loader">
+    <div className="loader-content animate-flicker">Please wait...</div>
+  </div>
+);
+
+export default Loader;
